test reservation et validation marche

+++ src/main/java/com/dto/LivreNombre.java
@@ -1,21 +1,21 @@
 package com.dto;
 
 public class LivreNombre {
-    private Integer id;
+    private Long id;
     private String titre;
     private Long nombre;
 
-    public LivreNombre(Integer id, String titre, Long nombre) {
+    public LivreNombre(Long id, String titre, Long nombre) {
         this.id = id;
         this.titre = titre;
         this.nombre = nombre;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
